refactor(agents): extract dynamic-agent types, identity, and RPC bridges
First step of moving the facet (sub-agent) machinery out of the Agent
god class into packages/agents/src/dynamic-agents/. Pure motion: the
moved types, identity helpers, and connection/reply bridges keep their
behavior; index.ts imports them under the old local names. Public
exports (SubAgentClass/SubAgentStub) are unchanged, now aliasing the
module's DynamicAgent* names. No wire- or storage-visible identifier
is renamed.
