+++ packages/agents/src/dynamic-agents/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Dynamic agents — facet-backed child agents, colocated with and
+ * supervised by their parent Agent.
+ *
+ * Everything exported from this barrel is **@experimental** and may
+ * change without a major version bump.
+ */
+
+export type { DynamicAgentClass, DynamicAgentStub } from "./types";
